refactor: reorganize project structure by moving background and content scripts to dedicated directories
