+++ src/sidepanel/components/ShareCapture.css
