+++ src/sidepanel/components/TagInjector.tsx
@@ -1,10 +1,10 @@
 import { AlertCircle, Send, Trash2 } from 'lucide-react';
 import { nanoid } from 'nanoid';
 import React, { useCallback, useState } from 'react';
-import { Logger } from '../lib/logger';
-import { ElementInfo } from '../types/types';
-import { formatElementTag } from '../utils/htmlTagFormatter';
-import { validateHtmlTag } from '../utils/htmlValidator';
+import { Logger } from '../../lib/logger';
+import { ElementInfo } from '../../types/types';
+import { formatElementTag } from '../../utils/htmlTagFormatter';
+import { validateHtmlTag } from '../../utils/htmlValidator';
 import { Card } from './common/Card';
 import { Tooltip } from './common/Tooltip';
