+++ src/sidepanel/components/common/Card.tsx
