+++ src/sidepanel/components/common/DOMTreeView.tsx
@@ -1,8 +1,8 @@
 // DOMTreeView.tsx
 import { ChevronDown, ChevronRight } from 'lucide-react';
 import React, { useState } from 'react';
-import { ElementInfo } from '../../types/types';
-import { formatElementTag } from '../../utils/htmlTagFormatter';
+import { ElementInfo } from '../../../types/types';
+import { formatElementTag } from '../../../utils/htmlTagFormatter';
 import { Tooltip } from './Tooltip';
 
 interface Props {
